+++ src/collective/volto/formsupport/scripts/cleansing.py
@@ -1,13 +1,12 @@
 # -*- coding: utf-8 -*-
 from collective.volto.formsupport.interfaces import IFormDataStore
-from datetime import datetime
-from datetime import timedelta
 from plone import api
 from zope.component import getMultiAdapter
 from zope.globalrequest import getRequest
-
+import sys
 import click
 import transaction
+from collective.volto.formsupport.restapi.services.form_data.form_data import FormData
 
 
 @click.command(
@@ -17,42 +16,55 @@ import transaction
         allow_extra_args=True,
     ),
 )
-@click.option("--dryrun/--no-dryrun", is_flag=True, default=True, help="--dryrun (default) simulate, --no-dryrun actually save the changes")
-@click.option("--days", type=int, default=6 * 30, help="default number of days of data retention (default 180)")
-def main(dryrun, days):
+@click.option(
+    "--dryrun/--no-dryrun",
+    is_flag=True,
+    default=True,
+    help="--dryrun (default) simulate, --no-dryrun actually save the changes",
+)
+def main(dryrun):
+    # import pdb;pdb.set_trace()
     if dryrun:
         print("CHECK ONLY")
     catalog = api.portal.get_tool("portal_catalog")
+    root_path = "/".join(api.portal.get().getPhysicalPath())
+    request = getRequest()
     if "blocks_type" in catalog.indexes():
-        brains = catalog(block_types="form")
+        brains = catalog.unrestrictedSearchResults(block_types="form", path=root_path)
     else:
         print("[WARN] This script is optimized for plone.volto >= 4.1.0")
-        brains = catalog()
+        brains = catalog.unrestrictedSearchResults(path=root_path)
     for brain in brains:
         obj = brain.getObject()
         blocks = getattr(obj, "blocks", None)
         if isinstance(blocks, dict):
-            for block in blocks.values():
+            for block_id, block in blocks.items():
                 if block.get("@type", "") != "form":
                     continue
                 if not block.get("store", False):
                     continue
-                if block.get("remove_data_after_days") in (-1, "-1"):
+                remove_data_after_days = int(block.get("remove_data_after_days") or 0)
+                # 0/None -> default value
+                # -1 -> don't remove
+                if remove_data_after_days <= 0:
+                    print(
+                        f"SKIP record cleanup from {brain.getPath()} block: {block_id}"
+                    )
                     continue
-                store = getMultiAdapter((obj, getRequest()), IFormDataStore)
-                expire_date = datetime.now() - timedelta(
-                    days=block.get("remove_data_after_days", days)
-                )
-                todelete = [
-                    record.intid
-                    for record in store.search()
-                    if record.attrs["date"] < expire_date
-                ]
-                if todelete:
-                    print(f"DELETE {len(todelete)} records from {brain.getPath()}")
-                    for intid in todelete:
-                        store.delete(intid)
-
+                data = FormData(obj, request, block_id)
+                store = getMultiAdapter((obj, request), IFormDataStore)
+                deleted = 0
+                for item in data.get_expired_items():
+                    store.delete(item["id"])
+                    deleted += 1
+                if deleted:
+                    print(
+                        f"[INFO] removed {deleted} records from {brain.getPath()} block: {block_id}"
+                    )
     if not dryrun:
         print("COMMIT")
         transaction.commit()
+
+
+if __name__ == "__main__":
+    sys.exit(main())